drivers/python/pyproject.toml
[project]
name = "bento-mdf"
version = "0.4.5"
description = "Python driver/validator for Bento Model Description Format"
authors = [
    { name="Mark A. Jensen", email = "mark.jensen@nih.gov"}
]
requires-python = ">=3.8"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License"
]

[project.urls]
"Homepage" = "https://github.com/CBIIT/bento-mdf"
"Bug Tracker" = "https://github.com/CBIIT/bento-mdf/issues"

[tool.poetry]
name = "bento-mdf"
version = "0.4.5"
description = "Python driver/validator for Bento Model Description Format"
authors = [
    "Mark A. Jensen <mark.jensen@nih.gov>"
]
license = "Apache 2.0"
readme = "README.md"
include = ["logs/log.ini"]

[tool.poetry.scripts]
"test-mdf.py" = { reference = "bin/test-mdf.py", type = "file" }
"load-mdf" = { reference = "bin/load-mdf.py", type = "file" }

[tool.poetry.dependencies]
python = "^3.8"
bento-meta = ">=0.0.54"
jsonschema = "^4.17.3"
pyyaml = ">=6"
delfick-project = "^0.7.9"
requests = "^2.28.2"
tqdm = "^4.64.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
jsonschema = "^4.17.3"
pyyaml = ">=6"
tqdm = "^4.64.1"
myst-nb = "^0.17.1"
sphinx-autoapi = "^2.0.1"
sphinx-rtd-theme = "^1.1.1"
jupyter = "^1.0.0"
python-semantic-release = "^7.33.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = "pyproject.toml:version"
branch = "master"
changelog_file = "CHANGELOG.md"
build_command = "poetry build"
dist_path = "dist/"
upload_to_release = true
upload_to_pypi = false
remove_dist = false
patch_without_tag = true
major_on_zero = false # while major version on 0.y.z, won't bump to 1.0.0
version_source = "tag" # temp?

[tool.pytest.ini_options]
filterwarnings = [
   "ignore::UserWarning",
   ]
   